requirements.txt

transformers
torch
streamlit
scikit-learn
faiss-cpu
